rfcfold: support reading from STDIN etc.
If the provided input file is not a regular file, but, e.g., STDIN,
write the contents to a temporary file for further processing.  This
allows the use of `-i /dev/stdin` to read input from another process
via pipe, if /dev/stdin is supported by, e.g., the OS or the shell.
Manually tested on GNU/Linux from within a Bash and a dash shell.