validate-all.sh: test reading data from stadard input
Users of rfcfold may use any method provided by their shell to specify
a file name that represents standard input.  Since the test script is
writen in Bash, and Bash provides the special filename "/dev/stdin"
to access standard input as a file, use that.

The tests use a couple of existing test cases to establish confidence
in the mechanism to read from standard input.