adjust sed scripts to conform to POSIX wording
This addresses two compatibility issues:

  1. The substitution part of an s command cannot portably contain
     the escape sequence '\n', but must use an embedded newline
     instead.

  2. Some sed commands may not be terminated with a semicolon, but
     must be terminated with a newline.

This is similar, but not identical, to sed script changes tested
successfully on macOS Catalina (10.15.7).  This still works with
GNU sed, but needs to be tested with macOS sed.  Reports of test
results with other "current" sed implementations are welcome, too.

After successfully testing sed script changes for macOS compatibility,
the warning when not using GNU sed should be removed.