recognize special name "/dev/stdin"
On Linux, where /dev/stdin is provided by the kernel, Bash just uses
this file.  Linux correctly reports it as a character special file,
thus "test -f" fails (i.e., /dev/stdin is not a regular file).

This seems not to be the case on macOS.  But since Bash emulates the
/dev/stdin behavior known from Linux whenever the OS does not natively
provide it, we can match on this name for the special case of reading
from standard input (aka STDIN).  Thus at least using rfcfold as any
part of a pipeline should be possible even on macOS.