change comment about gsed
Due to recent improvements, rfcfold no longer requires GNU sed.
Since GNU sed allows for a larger value for the folding column
than, e.g., the sed included in macOS, there is still a reason
to prefer GNU sed, if available.

Traditionally, GNU tools installed on a POSIX compatible system
use the prefix "g" to allow installation alongside the system
tools of the same name.  Thus look for a program named "gsed"
and use that instead of "sed," if found.

This logic is implemented in rfcfold already, but originally,
GNU sed was required, and the comment describing the choice
logic implied this.  Adjust the comment to better reflect the
current circumstances.