support reading from a pipe etc.
The intention of this commit is to enable use of `rfcfold` as any part of
a Unix *pipeline*.  Before, `rfcfold` could be used as the first program,
the source, in a pipeline by specifying `-o /dev/stdout`, but not as an
intermediate or final (sink) component.

If the provided input file is not a regular file, but, e.g., STDIN
reading from a pipe, write the contents to a temporary file for further
processing.  This allows the use of `-i /dev/stdin` to read input from
another process via pipe, if /dev/stdin is supported by, e.g., the OS
or the shell.  The Bash shell ensures the availability of `/dev/stdin`.

Manually tested on GNU/Linux from within a Bash and a dash shell.

Test this functionality by reading data from a pipe in `validate-all.sh`.
The added tests re-use existing test files and tests to verify that
reading data from a pipe does not negatively affect `rfcfold` operation.

This addresses GitHub issue #3.