test "unfolding is copy" with non-empty second line
The check for a non-empty second line looks as if it can be done
just once before looking for the presence of the first header
line.  But then "unfolding" a file that is not folded is not a
copy operation, unless the second line of the file is empty.

Add a test case for this condition.